add getProductsById
