+++ product-service/lambda/getProductsById.ts
@@ -1,0 +1,35 @@
+import {
+  APIGatewayProxyEvent,
+  APIGatewayProxyHandler,
+  APIGatewayProxyResult,
+} from "aws-lambda";
+import { PRODUCTS } from "./products";
+
+export const handler: APIGatewayProxyHandler = async (
+  event: APIGatewayProxyEvent
+): Promise<APIGatewayProxyResult> => {
+  const productId = event.pathParameters?.productId;
+  const product = PRODUCTS.find((product) => product.id === productId);
+  if (!product) {
+    return {
+      statusCode: 404,
+      headers: {
+        "Access-Control-Allow-Origin": "*",
+        "Access-Control-Allow-Headers": "Content-Type",
+        "Access-Control-Allow-Methods": "GET",
+        "Content-Type": "application/json",
+      },
+      body: JSON.stringify({ message: "Product not found" }),
+    };
+  }
+  return {
+    statusCode: 200,
+    headers: {
+      "Access-Control-Allow-Origin": "*",
+      "Access-Control-Allow-Headers": "Content-Type",
+      "Access-Control-Allow-Methods": "GET",
+      "Content-Type": "application/json",
+    },
+    body: JSON.stringify(product),
+  };
+};
