+++ product-service/lib/nodejs-aws-shop-react-be-stack.ts
@@ -1,10 +1,5 @@
 import { CfnOutput, Stack, StackProps } from "aws-cdk-lib";
-import {
-  Cors,
-  LambdaIntegration,
-  LambdaRestApi,
-  RestApi,
-} from "aws-cdk-lib/aws-apigateway";
+import { Cors, LambdaIntegration, RestApi } from "aws-cdk-lib/aws-apigateway";
 import { Code, Function, Runtime } from "aws-cdk-lib/aws-lambda";
 import { Construct } from "constructs";
 
@@ -12,17 +7,16 @@ export class NodejsAwsShopReactBeStack extends Stack {
   constructor(scope: Construct, id: string, props?: StackProps) {
     super(scope, id, props);
 
-    // define lambda resource
-    const hello = new Function(this, "HelloHandler", {
-      runtime: Runtime.NODEJS_18_X,
-      code: Code.fromAsset("lambda"),
-      handler: "hello.handler",
-    });
-
     const getProductsList = new Function(this, "GetProductsListHandler", {
       runtime: Runtime.NODEJS_18_X,
       code: Code.fromAsset("lambda"),
       handler: "getProductsList.handler",
+    });
+
+    const getProductsById = new Function(this, "GetProductsByIdHandler", {
+      runtime: Runtime.NODEJS_18_X,
+      code: Code.fromAsset("lambda"),
+      handler: "getProductsById.handler",
     });
 
     // define api gateway resource
@@ -38,7 +32,12 @@ export class NodejsAwsShopReactBeStack extends Stack {
         allowMethods: Cors.ALL_METHODS,
       },
     });
+
     productResource.addMethod("GET", new LambdaIntegration(getProductsList));
+
+    productResource
+      .addResource("{productId}")
+      .addMethod("GET", new LambdaIntegration(getProductsById));
 
     new CfnOutput(this, "GatewayUrl", { value: productResource.path });
   }
